Enable RS2 and RS3 testing in Helix (#226)
We were hitting an issue with test execution in Helix on RS2 and RS3 machines. Sometimes the test appx would fail to install. After debugging, the issue was traced to a problem with the appx package management system. Constantly installing and uninstalling the supporting framework packages (VCLibs and CoreRuntime) makes the issue more likely to be hit.

We work around the issue by manually installing these packages before running the test. This means that when the test app is uninstalled, the supporting packages do not get uninstalled with it, which avoids hitting this appx deployment issue.