Update TextCommandBarFlyout to handle selection of mixed format text (#261)
Fix for internal Bug 20352481: [SelectionFlyout] Selecting partially formatted text in RichEditBox extends the format to the other characters

There are a couple of issues here.

If you select text in a RichEditBox and bold it and then select unbolded text, when the TextCommandBarFlyout shows, it will bold it. The cause of this is that when the flyout shows it toggles the state of the toggle button to put it to its initial state (off). This causes the button's action to trigger which uses the FormatEffect::Toggle value to switch the bold state of the text. The result is that just selecting the unbolded text causes it to become bold. The fix here is to disable the toggle button actions until their initial state is set.

A second issue is with how the TextCommandBarFlyout handles mixed format text (e.g. a selection that contains bold and non-bold text). The behavior of FormatEffect::Toggle on mixed format text depends on the exact nature of the mixed formatting - if the selection starts with bold text, toggle will result in the selection being fully unbolded. If the selection starts with non-bold text, toggle will result in the selection being fully bolded. As a result, it is possible to get the bold state of the text to be the inverse of what the toggle button indicates. The fix is not to use FormatEffect::Toggle and instead explicitly set the FormatEffect to On or Off. This brings Italic and Bold in line with what Underline was doing (since there is no Toggle for Underline).

A third fix was to make the behavior of mixed formatted selections with Underline match the equivalent behavior of Bold and Italic. "Undefined" (i.e. mixed) is now treated the same as "Off". A mixed format selection is consistently treated as being in the "unformatted" state (i.e. the corresponding toggle button is Off). Clicking the toggle button will apply the formatting to the full selection (e.g. make the entire selection bold). Clicking a second time will remove the formatting from the full selection.

This PR also re-enables some tests which had been left disabled despite the referenced bug having been closed (likely due to a bad merge somewhere).