ColorSpectrum fractional width crash fix (#347)
We previously fixed an issue where ColorPicker would crash if its spectrum had a fractional width, but there were a couple other places as well where we were truncating rather than rounding. These changes fix those remaining issues and update the value used in the test to a value that exercises every circumstance where we crashed because of this.