Enabling test re-runs on failure to tell failing tests apart from unreliable tests (#709)
These changes enable the ability for us to re-run tests on failure, which will enable us to tell between test failures that are consistent versus test failures that are the result of periodic failures.  To do this, I've set things up so that we parse the te.wtl output file from TAEF to determine which tests failed in a given suite, and then we re-run those tests again to weed out tests that very intermittently fail.  After this new test run, we then run the tests that are still failing eight more times, and report a test as failing only if it fails at least 6/10 times, in order to further weed out the tests that fail every now and then.  This is more or less in line with what we do internally to differentiate between these two cases (with the exception of the fact that we run tests 50x there before concluding one way or another, which seemed excessive).

This does have one big shortcoming, which is that we're re-running tests on the same machine where they failed before, meaning that if there's anything systematically wrong with that machine, then we'll erroneously report a failure when there wasn't any problem with our code.  However, these events are rare, and this will definitely improve things over how things are right now.  Once we have our own Helix queue, we can re-evaluate doing a second test run instead of re-running tests on the same Helix machine.

We do need to do our own processing of test results to get the granular test results to be reported as part of the Azure DevOps build, as there's currently no way to report those results straight from Helix.  After talking with a Helix dev, I've [filed an issue on them](https://github.com/dotnet/arcade/issues/2773) to address that, and have added a TODO in our code to remind us to change things to make use of that improvement if/when it's available.

To test these changes, I created three dummy tests, one that always passes, one that fails 50% of the time, and one that fails 100% of the time.  I did a number of test runs, and in each case, we only reported the third of those three as failing.  I checked the logs and verified that we were indeed re-running the second test when it failed, and correctly logging that it had passed on re-run.

As part of implementing these changes, I moved the code to parse a te.wtl file into test pass objects into the TestPass class, since I needed to re-use that functionality for the purposes of extracting failed tests from the te.wtl file; it was previously only being used to output the testResults.xml file based on a single te.wtl file.