Enable file version info in build (#745)
When we switched over from Package ES builds to regular Azure DevOps builds, we lost some functionality in Package ES that was automatically setting the version info in the built dll.

This version info is used to determine what version of WinUI a given app is using.

This PR updates the build to set the version information.

When building locally, the version information is taken from version.props (currently defined to be 2.2.0.0). When building in Azure Pipelines, the third and forth component of this version is overridden based on the date (YY, MM, DD) and the pipeline's revision number (RRR) so that we get version 2.2.YYMM.DDRRR
This closely matches the nupkg which is versioned as:
2.2.YYMMDDRRR