Consolidating WinUI-Public-MUX-PR and WinUI-Public-Tests
Now that we're handling unreliable tests as separate from actual test failures, we can finally merge these two jobs into a single job and have only a single job as a PR gate. The only differences between these two jobs is that the latter also builds x86 release, and that the latter runs tests, so I've added those two steps to the former job. Once this is checked in, then we can disable and delete the latter job, as it'll be fully redundant.