Collapse header if AlwaysShowHeader is false and  PaneDisplayMode is Top (#1095)
* Collapse header if AlwaysShowHeader is false and  PaneDisplayMode is Top

When updating the header visibility, we were ignoring the AlwaysShowHeader property if the DisplayMode was minimal
The problem is that the PaneDisplayMode is Top, the displayMode is automatically Minimal, therefore we were also
ignoring AlwaysShowHeader property in this scenario. This commit fixes it by checking alto if the navigation is Top
in addition to checking if DisplayMode is Minimal.

Fix #757

* Add additional checks on AlwaysShowHeaderTest

When the test is not running with RS4 style, we change the display mode to Top
and check if the Header is still showing. Then we set AlwaysShowHeader to false
and check if the Header disappeared.

Fix #757

* commit suggestions from @licanhua

Fix a comment on the code according to @licanhua suggestions

#757