Changing the job attempt count check in OutputTestResults.ps1 to only function for PRs (#1115)
In my last change I had us make sure that we're only considering the latest test run in cases where a PR has multiple test runs associated with the same build. However, that caused problems in CI builds where we have multiple test runs in the same build definition, and Azure DevOps doesn't seem to provide an obvious way to tell between the two cases.

Since CI builds are the only builds that have multiple test runs in the same build definition, and since PR builds are the only case where we care about change authors scheduling multiple test runs in a row, I've changed things so we only do this behavior for PR builds.

If in the future we change the PR build definition to include multiple test runs, we'll need to revisit this, but for now this should suffice. I've added a comment indicating as much.