Fixing repeated entries in Microsoft.UI.Xaml.vcxproj.filters
## Description
Having $(OutDir) references in <PageRequiringCustomCompilation> in the root of Microsoft.UI.Xaml.vcxproj was causing references to that to get repeatedly added to Microsoft.UI.Xaml.vcxproj.filters.  We need to instead add those items in the targets where they're created to avoid this.

## Motivation and Context
Fixes #821

## How Has This Been Tested?
Did a full build and test run, verified that everything still passes as normal.