Merge feature/cbs to main (#5261)
* Merged PR 5977323: Fixes to enable CI and Release Pipelines to work reliably in internal WinUI2 repo

A few fixes:
Disabling unreliable test: [Test MUXControls\.ReleaseTest\.RepeaterNoCrashTest is unreliable \#4899](https://github.com/microsoft/microsoft-ui-xaml/issues/4899)

XamlIslandsTestApp was failing to install on test machines due to TargetPlatformMinVersion being tied to the version of SDK. This was working in the public repo, since we build with the 19H1 sdk. But in the internal repo we build with an internal sdk, so the app was failing to install on 19H1 machines.

WinUI-CI was using a 'taefQuery' parameter to skip some tests (NugetPkgTestsOnly='True') when run as a framework package test app. However, this was not being plumbed through for the Release Pipeline, which resulted in test failures. The yml is refactored a bit to avoid this problem.

Here are CI and Release builds passing with these fixes:
[\#WinUI2\-Internal\-CI\_2104\.23005 test fixes](https://microsoft.visualstudio.com/WinUI/_build/results?buildId=33778548&view=results)

[\#WinUI2\-Internal\-Release\_2104\.23002 test fixes](https://microsoft.visualstudio.com/WinUI/_build/results?buildId=33778555&view=results)

* Merged PR 5978459: Fix signing for Framework Package

There is an issue in the Release Pipeline where it was creating a framework package that contained unsigned dlls. This is because we sign the dlls in a Job after the main Build Job where the framework package is created.

Normally we work around this by re-creating the Framework Packages in the CreateNugetPackage job. However there was an error in the condition parameters.signConfig got renamed to parameters.signOutput, so we were skipping this step.

Fixing the condition introduced a new issue: we are building against the Internal SDK. So we needed that same SDK to create the framework package. So we install the required SDK on the machine.

The overall workflow can probably be simplified quite a bit. But I am going to make the smaller fix for now rather than trying to refactor the yml too much.

* Initial mica implementation

* Fix crash

* BaseLayerBrush

* Go back to 21330

* Updates for latest sdk

* Updates for merging

* update name

* huh

* bump to 1.0.5

* const

* Fix for multiple brushes

* Undo changes to sdk nuspec

* Remove unneeded namespace

* Merged PR 5987528: Create and Push CBS VPacks in Release Pipeline

#Create and push CBS VPack
This used to be a manual process. Instead we now do this as part of the Release Pipeline. Each time we run the Release Pipeline, we will create a CBS VPack and push it. The new VPack will not be used by anything until the appropriate manifests get updated to point to the new version. Doing this and creating a PR into the os repo remains a manual process.

##CreateCBSVPack.ps1
The creation of the VPacks is done by CreateCBSVPack.ps1. This is a port of:
`shellcommon\shell\Windows.UserExperience\Mbs\Microsoft-UI-Xaml-CBS\PublishMux.ps1`
from the OS repo. This script was run manually by a dev whenever they wanted to create a new set of VPacks from a drop of WinUI. The OS repo script did two things:
1. Create the CBS VPacks
2. Update the manifests in the OS repo to point to the new version of the VPacks.

Part 1 is now done by the Pipeline in CreateCBSVPack.ps1. Part 2 still needs to be done manually by a dev.

##Re-merge winmd
One of the steps in creating the CBS VPack is that we need to re-merge our winmd against the razzle metadata instead of against the Windows SDK metadata. This can be done by doing a VPack Pull on the Windows Publics which contains this metadata. However, the Windows publics is about 20 GB. So it takes about 20 minutes to do the VPack Pull on the build machine. We only need a single file from the publics, so this is very wasteful. There is not a way to selectively pull files from a VPack. So instead, I have packaged that single file (windows.winmd) into a nupkg (Microsoft.Internal.WinUI.WindowsPublicsWinmd) that is pushed to an internal feed (same one as used for the Internal SDK package). This should not need to be updated very often (if ever).

#Simplify signing and framework package workflow

The previous workflow of creating a framework package with signed dlls was convoluted. The Build Job would create the unsigned product dlls and framework package (containing the unsigned dlls).
The SignBinariesAndPublishSymbols Job would then sign the dlls and upload them to the drop to overwrite the unsigned dlls.
Then the CreateNugetPackage Job would then rebuild the framework package (which required installing the Internal Windows SDK) so that it contained the signed dlls and then upload that to the drop to overwrite the previous version.

The new workflow is simpler:
The Build Job builds the product dlls, signs them and packages them in a framework package. There is no re-building the framework package and no overwriting files on the drop.
The Build Job is also updated to publish the symbols so that we can get rid of SignBinariesAndPublishSymbols entirely.

* Change MicaController to be internal

* Missed some things

* CR feedback

* Fix warning

* Mica API review

* Backdrop tests and other refactoring

* Test page and test apis

* CR feedback

* Update test

* CR feedback and make the test actually verify stuff

* Add missing [webhosthidden]

* Fix test on RS2

* Add listener for ColorValuesChanged and reduce test on RS2

* Default tosystem theme if there's no window root when we set up mica

* add back version preview property until dependencies have removed using it in favor of ControlsResourcesVersion

* Fixes for ~MicaController getting called during cleanup

* WIP

* Don't animate between solid colors

* Revert unrelated changes

* Move back to public SDK

* Specify namespace to resolve ambiguity when compiling against insider sdk

Co-authored-by: Jevan Saks <jevansa@microsoft.com>
Co-authored-by: Ranjesh Jaganathan <ranjeshj@microsoft.com>
Co-authored-by: Ranjesh Jaganathan <28935693+ranjeshj@users.noreply.github.com>