Update framework package versioning to avoid version number size constraints
Our current versioning system for the framework package is soon going to run into the fact that framework package version numbers aren't allowed to exceed 65535. To avoid that, we'll move what was the minor version to be the major version of the framework package, which ensures that no portion of the version will be larger than that value. Since the actual WinUI version is part of the framework package name, this shouldn't create confusion regarding what NuGet package corresponds with what framework package. So, for example, the current framework package full name is this:

Microsoft.UI.Xaml.2.6_2.62107.6002.0_x64__8wekyb3d8bbwe

This change will instead make it look like this:

Microsoft.UI.Xaml.2.6_6.2107.6002.0_x64__8wekyb3d8bbwe