Backport WebView2 to WinUI2 (#5708)
This is a backport of the WinUI3 control and tests into WinUI2. Notable differences from the experience you get in WinUI3 via WinAppSDK are:

* This change adds a nuspec dependency from Microsoft.UI.Xaml onto WebView2's nupkg. In WinAppSDK the winrt implementation dlls are carried in the WinAppSDK runtime. We may change to do that as a follow-up.
* In WinUI3 the Focus related methods were moved so that WebView2 can derive from FrameworkElement. We don't have that in OS XAML so WebView2 control derives from Control instead. 🤷

WinUI2 also compiles and supports Arm32, but WebView2 package does not because there is no WebView2 runtime for Arm32. I've asked the WebView2 folks to consider adding at least the type information (with stub implementation), but I don't know if/when that will happen. As such I have made the build conditional so that the Arm32 version does not include WebView2 control. Consumers will need to write conditional code as well if they support Arm32 and want to use WebView2.

Fixes #5610