Changing CommandBarFlyout to not take keyboard focus when opened in transient mode (#6844)
CommandBarFlyout was changed to unconditionally give keyboard focus to the first focusable command when it opens. This breaks the assumption when opening in transient mode that the flyout won't take keyboard focus in that mode. We should only give keyboard focus when opening in standard mode.

Also added a test to make sure that this doesn't regress in the future.