Fixing SplitButtonAutomationPeer's Invoke implementation (#6918)
SplitButtonAutomationPeer's Invoke implementation currently just directly calls the Click event handler, which misses the fact that we also have the option of assigning a command to execute.  Instead of that, I've made it so we just forward the Invoke implementation to the primary button's automation peer, which will ensure that we do everything a button should do in this case.

I also edited the test case using the SplitButton with a command to use Invoke() to test that case as well.